feat(types): Support CDF ballot definition in safeParseElection
